Blender addon script and documentation (#1218)
* Blender addon script and documentation

First version of blender addon python script located in scripts/blender and new documentation for the addon. The Blender addon is for importing or exporting a camera path for visual effects and compositing.

* Updated formatting of Blender script

Updated the formatting based on the script tests for the blender addon script.

* Updated Blender addon to fix FOV and alignment

Updated how FOV is calculated for both blender cam to json and json to blender cam. The FOV is calculated based on the aspect ratio, angle of view, and sensor fit and size.

* Updated Blender addon documentation with examples

Updated the Blender add-on documentation with a few example gif composite animations

* Updated Blender addon to match frame rate and fixed formatting

Updated the Blender add-on script and docs to match the Blender frame rate. Also the formatting should be fixed. There are also now error message popups instead of console prints for invalid input in the add-on.

* Updated the blender addon for formatting

Updated pylint formatting and comments