feat: Update download and training commands for dozer data
The commit updates the download and training commands for the dozer data in the `pixi.toml` file. It renames the `download-sample-data` command to `download-dozer-data` and modifies the `train-example-splat` and `train-example-nerf` commands to include the dozer data as a dependency.

This makes it easy to run an example training for both nerfacto and splatfacto