made Dockerfile singularity/apptainer compatible
On HPC systems, singularity/apptainer is used instead of docker as the container solution.

The current dockerfile does not work under singularity, since nerfstudio is installed under /home/user, which is not accessible when the container is run in singularity/apptainer. The problem is that under singularity, the user inside the container has the same rights as outside the container and thus cannot access the contents of the folder belonging to another user. This makes singularity containers secure to use in a multi-user environment.

The proposed changes install nerfstudio in the system environment, making it available to all users.

I have tested this container and it works. The container is available on hub.docker.com/thawn/nerfstudio

closes #2952