Fix update of default values in ParallelDataManager
Fixed instantiation of ParallelDataManager to avoid modifying the default values of exclude_batch_keys_from_device.

In previous implementation, if masks_on_gpu is true, values are removed from the array. However, the variable contains a reference to the default array in InputDataset. Therefore, the default array is modified, and next instantiations of any InputDataset will have a truncated array without the removed value. Next ParallelDataManager instantiation will then fail to remove the value from the array (because the array does not have it anymore) and will raise an exception aborting everything.